chore: move imports to top of Strategies.py and remove circular dependency
