+++ code/backend/batch/utilities/helpers/DocumentChunkingHelper.py
@@ -1,10 +1,8 @@
 from typing import List
 from langchain.docstore.document import Document
-from ..document_chunking.Strategies import (
-    get_document_chunker,
-    ChunkingSettings,
-    ChunkingStrategy,
-)
+
+from ..document_chunking.ChunkingStrategy import ChunkingSettings, ChunkingStrategy
+from ..document_chunking.Strategies import get_document_chunker
 
 __all__ = ["ChunkingStrategy"]
